Update HA support to include GlusterFS
To eliminate the requirement for a separate HA shared filesystem, this
update adds support for using a glusterfs replicated volume as the
shared storage between the nodes.  This means that NF can convert
three less-reliable nodes into a cluster with hot-standby.

Since all buffered data and configuration are stored either in redis
or on the shared volume, all settings and buffered data will persist
even if a node is totally unrecoverable.